Mongo Atlas Basic Connection Established | Logger Added
Within this commit: Mongoose was adopted as a means of connecting a Mongo Atlas Cluster to the Express application. A basic 'user' model was also defined to showcase some data to be modelled for InterAct. Several other utility libraries such as Pino and dayJS where used to create a simple low-overhead logger, as the native console library blocks simple IO/threads