User Role Middleware Validation
Within this commit: A 'validateRole' middleware function was implemented to handle user permissions of resources on the application. This was implemented by saving the user's role on login to a global variable, which is then accessed within the middleware, and compared to the explicit role given to the function parameter.