Code Review #1 - Role enums, DB operation improvements & Cleanup
Within this commit: This commit focused on reviewing the implemented code and making adjustments where necessary; this included, but was not limited to, converting the hardcoded strings for each role, into an enum-like helper for consistency, while also creating a type 'ValueOf' as a measure of extracting our keys from our 'Roles' enum and adding them to a string array to use as either 'youth_worker' or 'student' as a string value. Mongoose database operations where also improved, by looking up a single document in the DB for user login & register actions using 'findOne()', rather than a list of documents with 'find()' - this significantly improved lookup performance.